+++ webapp/src/app/(admin-portal)/(form-page-layout)/[workspace_name]/dashboard/forms/[form_id]/deletion-requests/page.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import FormResponsesTable from '@Components/datatable/form-responses';
+import ResponseSegments from '@app/components/form/response-segments';
 import { selectForm } from '@app/store/forms/slice';
 import { useAppSelector } from '@app/store/hooks';
 import { selectWorkspace } from '@app/store/workspaces/slice';
@@ -10,6 +11,9 @@ export default function Page() {
     const workspace = useAppSelector(selectWorkspace);
 
     return (
-        <FormResponsesTable props={{ formId: form.formId, workspace, requestForDeletion: true }} />
+        <div className="mt-4 px-4 md:px-10 lg:px-28">
+            <ResponseSegments />
+            <FormResponsesTable props={{ formId: form.formId, workspace, requestForDeletion: true }} />
+        </div>
     );
 }